feature: breakout restrictions package
Breaking out the modules is a way to reduce the overhead of the project.
And allow users to only use whichever module they actually require.

This breakout will be used from the top level `bio_ex` package.

For a full commit history of the files, you can refer to
`https://github.com/bio-ex/bio_ex`.