feature: add `all/0` and updated docs
In order to effectively use the restriction enzymes, you basically need
a way to access them. There are two ways, when you know exactly the
enzyme you want, and when you have a function for filtering to what you
want. We only supported knowing exactly what you wanted.

This addresses that need by providing a way to list all the enzymes from
the module. This may not be the most efficient implementation, but it
appears to work well enough. The biggest downside that I can see is that
I had to create atoms for all the names of the enzymes. Which is a lot
of atoms.

Otherwise, I think that everything seems to work well.